Organize layers and fix chat identity
