+++ ChatApp/Data/ApplicationDbContext.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
+using ChatApp.Models;
 
 namespace ChatApp.Data
 {
