+++ ChatApp/Pages/Shared/_LoginPartial.cshtml
@@ -1,5 +1,5 @@
 @using Microsoft.AspNetCore.Identity
-@using ChatApp.Data
+@using ChatApp.Models
 @inject SignInManager<ApplicationUser> SignInManager
 @inject UserManager<ApplicationUser> UserManager
