+++ ChatApp/_Imports.razor
@@ -9,4 +9,4 @@
 @using Microsoft.JSInterop
 @using ChatApp
 @using ChatApp.Shared
-@using ChatApp.Data
+@using ChatApp.Models
